requirements.txt
Pillow
tkinterdnd2
pandas
openpyxl
pyinstaller
